graphs disabled for a now; added updating email section in profile
