+++ resources/js/Pages/Components/StyledButton.vue
@@ -7,7 +7,7 @@ defineProps({
 <template>
   <button
     :disabled="isProcessing"
-    class="mx-1 rounded-md bg-brown px-2.5 py-1 text-neutral-50 shadow-paper transition-all duration-300 hover:bg-brown/80 hover:shadow-none disabled:bg-brown/50"
+    class="mx-1 rounded-md bg-brown px-2.5 py-1 text-neutral-50 shadow-paper transition-all duration-300 hover:bg-brown/80 hover:shadow-none disabled:bg-brown/50 disabled:shadow-none"
   >
     <slot />
   </button>
