--- resources/js/types/inertia.d.ts
@@ -1,16 +1,0 @@
-export {};
-declare global {
-  export namespace inertia {
-    export interface Props {
-      user: {
-        id: number;
-        name: string;
-        email: string;
-        created_at: Date;
-        updated_at: Date;
-      };
-      errorBags: unknown;
-      errors: unknown;
-    }
-  }
-}
